requirements.txt
google-auth-oauthlib==0.4.1
google-api-python-client==1.7.11
youtube_dl==2020.1.24
requests==2.22.0
spotipy
pandas
apple-music-python
pyjwt
cryptography